requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.3
django-rest-framework==0.1.0
djangorestframework==3.10.1
idna==2.8
python-memcached==1.59
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
